feat: POST /user, SessionStorage

+++ app/src/main/java/com/towich/cosmicintrigue/ui/viewmodel/VoteViewModel.kt
@@ -32,7 +32,7 @@ class VoteViewModel(): ViewModel() {
             }
         }.start()
     }
-    fun getUserId():Int{
+    fun getUserId():Long{
         TODO("Возвращает id пользователя")
     }
 }
